dooh_ai/pyproject.toml
[project]
name = "dooh-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.35.3",
    "onnx>=1.19.0",
    "onnxconverter-common>=1.16.0",
    "onnxruntime>=1.23.1",
    "pillow>=11.3.0",
    "pillow-heif>=1.1.1",
    "ultralytics>=8.3.207",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
